+++ CHANGELOG.md
@@ -5,7 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [1.0.0] / 2023-12-11
-- First Release
+- Create `RevitProgressBar` and `RevitProgressBarUtils`
+- Add `StatusBarController`
+- Add `RevitRibbonController`
+- Add `ApplicationUtils`
+- Add `BalloonUtils`
 - Use `Form` with `DoEvent` (`DoEvent` is faster.)
 - Disable `RevitRibbonControl` when `RevitProgressBar` running.